Minor improvements: warnings, code style, etc

+++ ice-hockey-rink/utils.h
@@ -5,6 +5,6 @@
 
 // Functions defined in utils.cpp
 GLint euclidean_distance(GLint x0, GLint y0, GLint x1, GLint y1);
-void setPixel(GLint x, GLint y, GLint point_size, GLfloat c1, GLfloat c2, GLfloat c3);
+void setPixel(GLint x, GLint y, GLint point_size, GLdouble c1, GLdouble c2, GLdouble c3);
 
 #endif // UTILS_H
